backend/requirements.txt
alembic==1.12.0
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cloudinary
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-RESTful
Flask-SQLAlchemy
Flask-WTF
greenlet
gunicorn
idna
itsdangerous
Jinja2
Mako
MarkupSafe
packaging
Pillow
psycopg2
pycountry
PyJWT
python-slugify
pytz
requests
six
SQLAlchemy
text-unidecode
typing_extensions
urllib3
Werkzeug
WTForms
